RTC:Radxa's RTC hym8563 support second-level wake-up.
For the hym8653, use the TF function to configure timed wake-up when the wake-up time is less than 256 seconds, and use the AF function for wake-up when the wake-up time is more than 256 seconds.
To prevent sleep failure caused by an excessively short wake-up time, set the minimum wake-up time to no less than 3 seconds.